requirements.txt
aiohttp==3.11.11
aiomysql==0.2.0
bcrypt==4.2.1
fastapi==0.115.6
PyJWT==2.10.1
python-dotenv==1.0.1
uvicorn==0.34.0
